# Improve autograd profiler performance