@apaszke 2018-09-17
@pytorchbot retest this please